pyproject.toml
[project]
name = "shop-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai==0.5.0",
    "pypdf==4.2.0",
    "python-dotenv==0.20.0",
    "uvicorn>=0.35.0",
]
